Main: Move Main.hh to main.cc
main.cc contains the main() function. It should contain actual code that
is related to running j4dd.
